# CICLO + FINAL/CicloPlus_AgroCompliance_AiStudio_Package_v1.0.1/openapi/agro-compliance-openapi.yaml
openapi: 3.0.3
info:
  title: Agro Compliance Gates API
  version: 1.0.1
  description: 'API para controle de rastreabilidade de gado por lote/animal, evidências
    (RFID, GPS, balança, câmera), travas automáticas por divergência e liberação condicionada
    por ciclo/conferência e mudança de estágio.

    '
servers:
- url: http://localhost:3000
...
components:
  schemas:
...
    CreateMotherRequest:
      type: object
      required:
      - orgId
      - orgCode
      properties:
        orgId:
          type: string
        orgCode:
          type: string
          minLength: 2
          maxLength: 3
          description: Codigo curto da organizacao (2 a 3 caracteres).
    Mother:
      type: object
      properties:
        orgId:
          type: string
        motherId:
          type: string
        motherCode:
          type: string
          description: Codigo curto ORG-SEQ-DC
        seq:
          type: integer
        childSeqCounter:
          type: integer
        status:
          type: string
          enum:
          - ATIVA
          - BAIXADA
...
    OpenCycleRequest:
      type: object
      required:
      - orgId
      - periodStart
      - periodEnd
      properties:
        orgId:
          type: string
        periodStart:
          type: string
          format: date
        periodEnd:
          type: string
          format: date
    Cycle:
      type: object
      properties:
        orgId:
          type: string
        cycleId:
          type: string
        status:
          type: string
          enum:
          - OPEN
          - PENDING_CONFERENCE
          - CLOSED
        periodStart:
          type: string
          format: date
        periodEnd:
          type: string
          format: date
        inventory:
          type: object
          properties:
            completed:
              type: boolean
        conference:
          type: object
          properties:
            conferenceId:
              type: string
              nullable: true
            status:
              type: string
              enum:
              - PENDING
              - APPROVED
              - REJECTED
            result:
              type: string
              enum:
              - OK
              - INCONSISTENT
              - MISSING_DATA
              nullable: true
    GateRequest:
      type: object
      required:
      - orgId
      - op
      properties:
        orgId:
          type: string
        op:
          type: string
          enum:
          - ALIMENTACAO
          - MEDICAMENTO
          - TRANSFERENCIA
          - NF_VENDA
    GateResponse:
      type: object
      properties:
        allowed:
          type: boolean
        reason:
          type: string
          nullable: true
...
paths:
...
  /mothers:
    post:
      tags:
      - Matrizes
      summary: Criar matriz (mae) com codigo curto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMotherRequest'
      responses:
        '200':
          description: Matriz criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mother'
...
  /cycles/open:
    post:
      tags:
      - Ciclos
      summary: Abrir ciclo operacional
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpenCycleRequest'
      responses:
        '200':
          description: Ciclo aberto
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cycle'
...
  /lotes/{loteId}/gate:
    post:
      tags:
      - Operacoes
      summary: Validar se uma operacao critica esta liberada para o lote
      parameters:
      - in: path
        name: loteId
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GateRequest'
      responses:
        '200':
          description: Resultado do gate
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GateResponse'
...
  /consumo/alimentacao:
    post:
      summary: Consumir alimentação
      responses:
        '200':
          description: OK
        '423':
          description: Bloqueado
  /consumo/medicamentos:
    post:
      summary: Consumir medicamentos
      responses:
        '200':
          description: OK
        '423':
          description: Bloqueado